+++ template_command.js
@@ -1,0 +1,21 @@
+module.exports = {
+	name: 'command',
+    description: 'Description',
+    // shortDescription: 'Short Description',
+    // args: true,
+	// aliases: ['alias'],
+    // usage: '<params...>',
+    // uses: [
+    //     "",
+    // ].join('\n'),
+    // arguments: [
+    //     "",
+    // ].join('\n'),
+    // userPermissions: [],
+    // botPermissions: [],
+    // guildOnly: true,
+    // adminOnly: true,
+	async execute(message, args, client) {
+        
+	},
+};